refactor(dsl): rename flow-related files and update import paths
Renamed files:
- flow.spec.ts -> dsl.spec.ts
- new-flow.ts -> dsl.ts
- Updated import statement in example-flow.ts to reflect new file name
